requirements.txt
Pillow>=9.0.0
customtkinter>=5.2.0
requests>=2.28.0
gspread>=5.7.0
google-auth>=2.17.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
oauth2client>=4.1.3
cryptography>=41.0.0
Flask>=2.3.0
Flask-CORS>=4.0.0
